// from-app/server/models/user.ts
export const API_BASE_URL = 'https://jsonplaceholder.typicode.com';

export interface Geo {
    lat: string;
    lng: string;
}

export interface Address {
    street: string;
    suite: string;
    city: string;
    zipcode: string;
    geo: Geo;
}

export interface Company {
    name: string;
    catchPhrase: string;
    bs: string;
}

export interface User {
    id: number;
    name: string;
    username: string;
    email: string;
    address: Address;
    phone: string;
    website: string;
    company: Company;
}

export interface AddUser
    extends Omit<User, 'id' | 'username' | 'address' | 'phone' | 'website'> {}